# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fiteam API Documentation",
    "description": "",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/user/savecard": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "사용자의 성향검사 답변을 결과형태로 저장. 질문의 typeA:I, typeB:E 일 때 응답이 4라면-\u003E I:1, E:4",
        "description": "[{E:4, I:1}, {P:2, D:3}...] 이렇게 사용자 응답결과를 RequestBody로 주세요",
        "operationId": "saveTestResult",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestAnswerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/question/unauth/test-result": {
      "post": {
        "tags": [
          "character-question-controller"
        ],
        "summary": "로그인 없이 성향검사 결과 연산",
        "description": "질문 결과를 연산하여 캐릭터카드 ID와 4가지 성향 점수를 반환합니다.",
        "operationId": "saveTestResult_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTestAnswerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SaveTestResultResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/user/settings": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "마이페이지 내정보 조회",
        "description": "로그인한 사용자의 전화번호, 카카오ID, 직업, 전공, 소개, URL을 반환합니다.",
        "operationId": "getUserSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponseDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "user-controller"
        ],
        "summary": "마이페이지 내정보 설정 변경",
        "description": "전화번호, 카카오톡 ID, 직업, 전공, 소개, URL을 수정합니다. null로 넘어온 값은 변경하지 않습니다.",
        "operationId": "updateUserSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/accept/{groupId}": {
      "patch": {
        "tags": [
          "user-controller"
        ],
        "summary": "유저가 받은 그룹 초대에서 그룹 참여",
        "description": "현재 사용자가 그룹 초대를 수락합니다.",
        "operationId": "acceptInvitation",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SaveTestAnswerRequestDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "description": "질문별 답변 리스트\", example = \"[{\\\"E\\\":4,\\\"I\\\":1}, {\\\"P\\\":3,\"D\":2}]",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        }
      },
...
      "SaveTestResultResponseDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "integer",
            "format": "int32",
            "description": "판단된 캐릭터카드 ID",
            "example": 12
          },
          "numEI": {
            "type": "integer",
            "format": "int32",
            "description": "EI 성향 점수",
            "example": 7
          },
          "numPD": {
            "type": "integer",
            "format": "int32",
            "description": "PD 성향 점수",
            "example": 5
          },
          "numVA": {
            "type": "integer",
            "format": "int32",
            "description": "VA 성향 점수",
            "example": 4
          },
          "numCL": {
            "type": "integer",
            "format": "int32",
            "description": "CL 성향 점수",
            "example": 6
          }
        }
      },
...
      "UserSettingsRequestDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "전화번호",
            "example": "010-1234-5678"
          },
          "kakaoId": {
            "type": "string",
            "description": "카카오톡 ID",
            "example": "my_kakao_id"
          },
          "job": {
            "type": "string",
            "description": "직업",
            "example": "Frontend Developer"
          },
          "major": {
            "type": "string",
            "description": "전공",
            "example": "Computer Science"
          },
          "introduction": {
            "type": "string",
            "description": "자기소개",
            "example": "안녕하세요. 프론트엔드 개발자입니다."
          },
          "url": {
            "type": "string",
            "description": "웹사이트 URL (GitHub, Blog)",
            "example": "https://github.com/username"
          }
        }
      },
...
      "UserSettingsResponseDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "description": "사용자 이름",
            "example": "고양이"
          },
          "profileImgUrl": {
            "type": "string",
            "description": "프로필 이미지 URL",
            "example": "https://example.com/profile.jpg"
          },
          "phoneNumber": {
            "type": "string",
            "description": "전화번호",
            "example": "010-1234-5678"
          },
          "kakaoId": {
            "type": "string",
            "description": "카카오톡 ID",
            "example": "my_kakao_id"
          },
          "job": {
            "type": "string",
            "description": "직업",
            "example": "Frontend Developer"
          },
          "major": {
            "type": "string",
            "description": "전공",
            "example": "Computer Science"
          },
          "introduction": {
            "type": "string",
            "description": "자기소개",
            "example": "안녕하세요. 프론트엔드 개발자입니다."
          },
          "url": {
            "type": "string",
            "description": "웹사이트 URL (GitHub, Blog)",
            "example": "https://github.com/username"
          },
          "cardId1": {
            "type": "integer",
            "format": "int32",
            "description": "CharacterCard1 ID",
            "example": 1
          },
          "cardId2": {
            "type": "integer",
            "format": "int32",
            "description": "CharacterCard2 ID",
            "example": 2
          },
          "details": {
            "type": "string",
            "description": "캐릭터 카드 AI 분석결과",
            "example": "ISTP 유형이면서 ~~한 특징이 있는 사람입니다."
          },
          "numEI": {
            "type": "integer",
            "format": "int32",
            "description": "EI 성향 점수",
            "example": 7
          },
          "numPD": {
            "type": "integer",
            "format": "int32",
            "description": "PD 성향 점수",
            "example": 3
          },
          "numVA": {
            "type": "integer",
            "format": "int32",
            "description": "VA 성향 점수",
            "example": 5
          },
          "numCL": {
            "type": "integer",
            "format": "int32",
            "description": "CL 성향 점수",
            "example": 8
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "계정 생성일시",
            "example": "2025-05-01T12:00:00"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}